Change request dataset name using the dataset name from ReqMgr
There is a mismatch on one character for the primary dataset between McM
and ReqMgr. Normally, the output dataset from ReqMgr must follow the dataset name convention
registered in McM.

Hot fix for a consistency error discovered today with ReqMgr.
Due to the output dataset is now produced, replace the dataset name on McM to be able to continue
with other steps.