Full injection from scratch
This provides two test case scenarios to submit a root request from scratch and a
full MC sample from scratch (up to NanoAOD data tier) to CMS O&C ReqMgr2 application,
providing some kind of "end-to-end test case scenario".

Currently, to use this, it is required that the testing environment configures the McM application with the required credentials
to send requests to ReqMgr2, open sessions in SSH to access AFS folders, and that the deployment can reach resources inside
the CERN internal network. Because the last is not achievable with the container environment created in previous iterations,
just use the current McM development deployment.

However, to properly authenticate the requests and reach some databases, the test module has to be executed in CERN internal nodes
and the user executing this must authenticate the client using an OIDC Device Code approach.

Note the last is just a temporal workaround to execute the test and a
complete testing environment must be created