Limit database responses to 10MB
We need to protect against running out of memory.

This also improves error reporting when database errors are encountered,
by raising an exception to abort the current operation and allow callers
to report the error to the user. Only the /search endpoint handles the
errors properly, but it is also a risky one.