Filter workflows by last injection date
Use the latest submitted date (related to the last injection event) to
filter all the workflow names linked to a request. Parse dates related
to objects on ReqMgr2 and McM to UTC to perform the comparison. If the
last submitted date is not available in the history, return the received
list.