Limit infinite requests (#1159)
* Limit database responses to 10MB

We need to protect against running out of memory.

This also improves error reporting when database errors are encountered,
by raising an exception to abort the current operation and allow callers
to report the error to the user. Only the /search endpoint handles the
errors properly, but it is also a risky one.

* Limit the number of returned DB objects to 1000

It was previously 9999, but this limit was a bit too high and could kill
the server. Decrease it by one order of magnitude.

* Improve error messages

Co-authored-by: Geovanny González-Rodríguez <geovanny.gonzalez@cern.ch>

* Avoid raising errors in database operations

Instead of raising a `ValueError` that could be uncaught by other
endpoints or internal functions include an optional message on the
database response to indicate errors related to the maximum response
size for database queries.

---------

Co-authored-by: Louis Moureaux <louis.moureaux@cern.ch>