Tests for the main application entities (#1155)
* Avoid collisions for campaign `prepid`s

Increase the window for generating numbers

* Remove the usage of `Entity` and `EntityTest` classes

1. Use the `InjectRootRequest` and `InjectToNanoAOD` classes to create
mock entities to use on the assertions instead of repeating these tasks
per each entity on the McM application.

2. Move the `test_campaign` package one level above

* Test for `chained campaigns`

Reuse the root chained campaign created automatically by McM when a new
campaign is created. Because of the last, no test is provided for the
`create` operation.

* Test for `tags` entity

* Test for `settings` entity

* Test for `lists` entity

* Test for `users` entity

* Test for `invalidations` entity

* Test for `flows` entity

* Test for `chained_request` entity

* Test for `request` entity

* Test for `mccms` entity

* Set `config` inside the test class for `test_api_tools`

Use `setup_method` to set the McM instance for the tests instead of
using a global scope variable and singleton style.

* Change flag for tests targeting the current dev environment

Just check that the environment variable `ABLE_TO_ACCESS_INTERNAL_RESOURCES`
is set, instead of raising a skip test signal. Refactor the fixture
`mcm_client` to just skip the test if the intended target is the
development environment.

Update fixtures `injection_to_reqmgr` and `able_to_access_internal_resources`
to enable `autouse`, this eases integration with class-style tests to use
this fixtures in method and class scopes.

* Skip test cases that inject requests to ReqMgr2 by default

Decorate them using the fixture `injection_to_reqmgr`. Set the
environment variable `REQMGR2_ENABLED` to execute these test cases.
This variable confirms that the testing environment is prepared to
inject requests into this service.

* Reorder imports - isort

* Change email address for testing user

This address will be used on email notifications. Follow guidelines
described at: Anonymous Smtp Posting [1].

[1]: https://mailservices.docs.cern.ch/Miscellaneous/Anonymous_smtp/