Include support for `gssapi-with-mic` for SSH sessions
Enable both `SSH` executors to authenticate to the remote host using
Kerberos tickets via `gssapi-with-mic`.

Remember to install the optional packages `gssapi` and `pyasn1` to
enable `paramiko` to use this authentication strategy.

Use `gssapi_with_mic` dynamically

In case the protocol is available to use, enable its use dynamically via
environment variables.

This is useful to manage the authentication strategy for SSH in local
environments where Kerberos is not configured.