Include `CleanForceFlowList` script
Include the script in the repository and update it to use the PdmV REST
client.

This script is used for maintaining the `list_of_forceflow` in McM and
it is executed by the batch job: 'McM Lists Maintenance'. The original
version of the script has been located in:
'/afs/cern.ch/cms/PPD/PdmV/tools/McM-admins/CleanForceFlowList.py'
without any copy registered in the repository. Include it in the
`automatic_scripts` folder.