Update `validation_control` module
Fix integration issues with the `installer` module. Provide an old-style
`ssh_executor` for checking the existence and possibly creation of a folder.
Also, avoid decoding an email message as the object is already `<cls: str>`.

In 424f07c34fd481172ddc1bcbd66ec2c26fea587e, a change was introduced to
the `installer` module to avoid executing file operations over a local filesystem.
The last forced the application runtime environment to always mount
remote filesystems like AFS or EOS. The solution provided shifts this
responsibility to a remote session.