Include a new validation strategy
Include a strategy to run validation jobs based on the number of
output events and set dynamically the number of events to run depending
on the thread configuration. Refactor the validation procedure to set
dynamically different validation strategies for computing the number of
events, set the resources to use in a validation job and submit
different job configurations. Also, remove the constraint of always
submit the first validation job with one thread.

* Fix: Submit the next stage discarding DONE threads

Improve the condition for submitting the next scenarios, instead of
discarding always the minimum, discard those already in done.

There is a side effect if the `time_per_event` is remeasured/corrected by the
first validation attempt, it could be possible that validation scenarios
discarded in the first iteration appear in the `next stage` procedure,
if the minimum is discarded, a validation thread job could get repeated.

* Fix: Consider only the scenarios common to all requests

If a chained request is validated, consider only the thread scenarios
common to all the request included in the chained request.

Co-authored-by: Jean-Roch Vlimant <vlimant@cern.ch>
Signed-off-by: Geovanny González-Rodríguez <geovanny.gonzalez@cern.ch>