Include a new validation strategy
Include a strategy to run validation jobs based on the number of
output events and set dynamically the number of events to run depending
on the thread configuration. Refactor the validation procedure to set
dynamically different validation strategies for computing the number of
events, set the resources to use in a validation job and submit
different job configurations. Also, remove the constraint of always
submit the first validation job with one thread.

Co-authored-by: Jean-Roch Vlimant <vlimant@cern.ch>
Signed-off-by: Geovanny González-Rodríguez <geovanny.gonzalez@cern.ch>