Check on action parameter flag during chain submission
The current chain submission does not check if a chained request is
ready to be injected. This readiness is set by the
`action_parameter.flag` attribute assgined to `true` for an active chain, and
`false` for inactive ones.

Currently, if we have a root request linked to several chained request
and it is desired to submit just one of them, even if the flag is set to
`false` for the others, at the moment of submitting the root request it
takes all of them.

Co-authored-by: Geovanny González-Rodríguez <geovanny.gonzalez@cern.ch>